+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `FileUtil.createLocalTempDir()` helper method.
+
 ### [v0.08](https://github.com/realityforge/gir/tree/v0.08) (2019-03-25)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.07...v0.08)